l4-adk-mcp/pyproject.toml
[project]
name = "l4-adk-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.22.1",
    "mcp[cli]>=1.25.0",
]
